Move generated code into a separate folder, improve testing
