+++ inst/cpp_generation/parameter_types.hpp.in
@@ -2,9 +2,9 @@
 
 #pragma once
 
-#include "types.hpp"
-#include "state_space.hpp"
-#include "model_variants.hpp"
+#include "../intermediate_data.hpp"
+#include "../state_space.hpp"
+#include "../model_variants.hpp"
 
 namespace leapfrog {
