+++ inst/cpp_generation/state_types.hpp.in
@@ -2,10 +2,10 @@
 
 #pragma once
 
-#include "types.hpp"
-#include "parameters.hpp"
-#include "state_space.hpp"
-#include "model_variants.hpp"
+#include "../intermediate_data.hpp"
+#include "../parameters.hpp"
+#include "../state_space.hpp"
+#include "../model_variants.hpp"
 
 namespace leapfrog {
