+++ inst/include/general_demographic_projection.hpp
@@ -1,7 +1,7 @@
 #pragma once
 
-#include "types.hpp"
-#include "state_types.hpp"
+#include "intermediate_data.hpp"
+#include "generated/state_types.hpp"
 
 namespace leapfrog {
