+++ inst/include/hiv_demographic_projection.hpp
@@ -1,8 +1,8 @@
 #pragma once
 
 #include <iostream>
-#include "types.hpp"
-#include "state_types.hpp"
+#include "intermediate_data.hpp"
+#include "generated/state_types.hpp"
 
 namespace leapfrog {
