+++ scripts/generate
@@ -8,7 +8,7 @@ source(file.path(root, "R/util_assert.R"))
 source(file.path(root, "R/generate_cpp_validation.R"))
 source(file.path(root, "R/generate_cpp.R"))
 
-frogger_file <- function(..., mustWork = TRUE) {
+frogger_file <- function(..., mustWork = FALSE) {
   path <- file.path(root, "inst", ...)
   if (mustWork && !file.exists(path)) {
     stop(sprintf("File at path %s does not exist", path))
@@ -16,7 +16,7 @@ frogger_file <- function(..., mustWork = TRUE) {
   path
 }
 
-generate_input_interface(dest = frogger_file("include/model_input.hpp"))
-generate_output_interface(dest = frogger_file("/include/model_output.hpp"))
-generate_parameter_types(dest = frogger_file("/include/parameter_types.hpp"))
-generate_state_types(dest = frogger_file("/include/state_types.hpp"))
+generate_input_interface(dest = frogger_file("/include/generated/model_input.hpp"))
+generate_output_interface(dest = frogger_file("/include/generated/model_output.hpp"))
+generate_parameter_types(dest = frogger_file("/include/generated/parameter_types.hpp"))
+generate_state_types(dest = frogger_file("/include/generated/state_types.hpp"))
